add more features

+++ typeflow-aspnet/TypeFlow.Application/Services/User/Dto/FullUserData.cs
@@ -3,7 +3,7 @@
     public class FullUserData : UserData
     {
         public float Accuracy { get; set; }
-        public int TotalTests { get; set; }
+        public long TotalTests { get; set; }
         public int AverageCPM { get; set; }
         public int AverageWPM { get; set; }
     }
